+++ src/cli/commands/sets/update.rs
@@ -46,7 +46,8 @@ pub struct UpdateArgs {
 }
 
 impl CommandRunner for UpdateArgs {
-    // FIXME: The signer is not required here
+    const NEED_SIGNER: bool = false;
+
     async fn run(self, mut options: CliOptions) -> N34Result<()> {
         let naddrs = self.naddrs.flat_naddrs(&options.config.sets)?;
         let relays = self.relays.flat_relays(&options.config.sets)?;