+++ CHANGELOG.md
@@ -18,6 +18,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Refactor
 
 - Store the config in `CliOptions` instead of its path - by Awiteb
+- Improve commands running and make the signer optional for some commands - by Awiteb
 
 ## [0.2.0] - 2025-06-01